refactor compareExperiments resolver on query
