+++ app/schema.graphql
@@ -100,7 +100,6 @@ type Dataset implements Node {
   """
   experimentCount(datasetVersionId: GlobalID): Int!
   experiments(first: Int = 50, last: Int, after: String, before: String): ExperimentConnection!
-  compareExperiments(baselineExperimentId: GlobalID!, comparisonExperimentIds: [GlobalID!]!): [ExperimentComparison!]!
 }
 
 enum DatasetColumn {
@@ -936,7 +935,12 @@ type PromptResponse {
 
 type Query {
   projects(first: Int = 50, last: Int, after: String, before: String): ProjectConnection!
+<<<<<<< HEAD
   datasets(first: Int = 50, last: Int, after: String, before: String, sort: DatasetSort): DatasetConnection!
+=======
+  datasets(first: Int = 50, last: Int, after: String, before: String): DatasetConnection!
+  compareExperiments(baselineExperimentId: GlobalID!, comparisonExperimentIds: [GlobalID!]!): [ExperimentComparison!]!
+>>>>>>> eedd4401 (refactor compareExperiments resolver on query)
   functionality: Functionality!
   model: Model!
   node(id: GlobalID!): Node!
