+++ src/phoenix/server/api/queries.py
@@ -1,5 +1,5 @@
 from collections import defaultdict
-from typing import Dict, List, Optional, Set, Union
+from typing import DefaultDict, Dict, List, Optional, Set, Union
 
 import numpy as np
 import numpy.typing as npt
@@ -9,7 +9,7 @@ from sqlalchemy.orm import joinedload
 from strawberry import ID, UNSET
 from strawberry.relay import Connection, GlobalID, Node
 from strawberry.types import Info
-from typing_extensions import Annotated
+from typing_extensions import Annotated, TypeAlias
 
 from phoenix.db import models
 from phoenix.pointcloud.clustering import Hdbscan
@@ -33,11 +33,12 @@ from phoenix.server.api.types.EmbeddingDimension import (
 )
 from phoenix.server.api.types.Event import create_event_id, unpack_event_id
 from phoenix.server.api.types.Experiment import Experiment
+from phoenix.server.api.types.ExperimentComparison import ExperimentComparison, RunComparisonItem
 from phoenix.server.api.types.ExperimentRun import ExperimentRun, to_gql_experiment_run
 from phoenix.server.api.types.Functionality import Functionality
 from phoenix.server.api.types.InferencesRole import AncillaryInferencesRole, InferencesRole
 from phoenix.server.api.types.Model import Model
-from phoenix.server.api.types.node import from_global_id
+from phoenix.server.api.types.node import from_global_id, from_global_id_with_expected_type
 from phoenix.server.api.types.pagination import (
     ConnectionArgs,
     CursorString,
@@ -104,6 +105,96 @@ class Query:
         return connection_from_list(
             data=[to_gql_dataset(dataset) for dataset in datasets], args=args
         )
+
+    @strawberry.field
+    async def compare_experiments(
+        self,
+        info: Info[Context, None],
+        baseline_experiment_id: GlobalID,
+        comparison_experiment_ids: List[GlobalID],
+    ) -> List[ExperimentComparison]:
+        OrmExample = models.DatasetExample
+        OrmExperiment = models.Experiment
+        OrmRun: TypeAlias = models.ExperimentRun
+        OrmRevision = models.DatasetExampleRevision
+        OrmTrace = models.Trace
+
+        experiment_ids = [
+            from_global_id_with_expected_type(experiment_id, OrmExperiment.__name__)
+            for experiment_id in [baseline_experiment_id] + comparison_experiment_ids
+        ]
+
+        async with info.context.db() as session:
+            if (
+                version_id := await session.scalar(
+                    select(func.max(OrmExperiment.dataset_version_id)).where(
+                        OrmExperiment.id.in_(experiment_ids),
+                    )
+                )
+            ) is None:
+                raise ValueError
+
+            revision_ids = (
+                select(func.max(OrmRevision.id))
+                .where(OrmRevision.dataset_version_id <= version_id)
+                .group_by(OrmRevision.dataset_example_id)
+                .scalar_subquery()
+            )
+            examples = (
+                await session.scalars(
+                    select(OrmExample)
+                    .join(OrmRevision, OrmExample.id == OrmRevision.dataset_example_id)
+                    .where(
+                        and_(
+                            OrmRevision.id.in_(revision_ids),
+                            OrmRevision.revision_kind != "DELETE",
+                        )
+                    )
+                    .order_by(OrmRevision.dataset_example_id.desc())
+                )
+            ).all()
+
+            ExampleID: TypeAlias = int
+            ExperimentID: TypeAlias = int
+            runs: DefaultDict[ExampleID, DefaultDict[ExperimentID, List[OrmRun]]] = defaultdict(
+                lambda: defaultdict(list)
+            )
+
+            for run in await session.scalars(
+                select(OrmRun)
+                .where(
+                    and_(
+                        OrmRun.dataset_example_id.in_(example.id for example in examples),
+                        OrmRun.experiment_id.in_(experiment_ids),
+                    )
+                )
+                .options(joinedload(OrmRun.trace).load_only(OrmTrace.trace_id))
+            ):
+                runs[run.dataset_example_id][run.experiment_id].append(run)
+
+        experiment_comparisons = []
+        for example in examples:
+            run_comparison_items = []
+            for experiment_id in experiment_ids:
+                run_comparison_items.append(
+                    RunComparisonItem(
+                        experiment_id=GlobalID(Experiment.__name__, str(experiment_id)),
+                        runs=[
+                            to_gql_experiment_run(run) for run in runs[example.id][experiment_id]
+                        ],
+                    )
+                )
+            experiment_comparisons.append(
+                ExperimentComparison(
+                    example=DatasetExample(
+                        id_attr=example.id,
+                        created_at=example.created_at,
+                        version_id=version_id,
+                    ),
+                    run_comparison_items=run_comparison_items,
+                )
+            )
+        return experiment_comparisons
 
     @strawberry.field
     async def functionality(self, info: Info[Context, None]) -> "Functionality":
